requirements.txt
anyio==4.9.0
blinker==1.9.0
boto3==1.39.14
botocore==1.39.14
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.3
requests==2.32.4
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
telegram==0.0.1
urllib3==2.5.0
Werkzeug==3.1.3
